Added contails_tol function
contains_tol tests to see if a point is inside a box within a certain tolerance. This is essential if you are finding intersection points as it allows for you to eliminate any issues of a point on the box's face not being detected as inside due to floating point errors. The default tolerance is set to 1e-14 for this reason.